requirements.txt
# Yahoo Fantasy Sports API - using latest stable version
yfpy>=16.0.0,<17.0.0

# Data processing - stable versions
pandas==2.1.4
numpy==1.26.4

# Date/time handling
python-dateutil==2.8.2

# HTTP requests (for API calls and webhooks)
requests==2.31.0

# JSON handling and utilities
python-dotenv==1.0.1

# Logging
colorlog==6.8.2

# Email support for notifications
# Note: smtplib and email.mime are part of Python standard library

# Additional dependencies to ensure compatibility
setuptools>=65.0.0
wheel>=0.41.0
